+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project attempts to adhere to [Semantic Versioning](https://semver.org/
 
 - Validate that the New Architecture works in the example app.
   - Required a workaround for CMAKE due to long paths in Windows.
+- Validate compatibility with React Native 0.79 & Expo SDK 53.
 
 ## [0.8.0] - 2025-02-13